Move helpers from session to context
Remove a circular dependency between session and context, and limits the
usage of implicits.
