+++ ledger/ledger-api-test-tool/src/main/scala/com/daml/ledger/acceptance/infrastructure/LedgerServices.scala
@@ -1,0 +1,24 @@
+// Copyright (c) 2019 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package com.daml.ledger.acceptance.infrastructure
+
+import com.digitalasset.ledger.api.v1.admin.party_management_service.PartyManagementServiceGrpc
+import com.digitalasset.ledger.api.v1.admin.party_management_service.PartyManagementServiceGrpc.PartyManagementService
+import com.digitalasset.ledger.api.v1.command_service.CommandServiceGrpc
+import com.digitalasset.ledger.api.v1.command_service.CommandServiceGrpc.CommandService
+import com.digitalasset.ledger.api.v1.ledger_identity_service.LedgerIdentityServiceGrpc
+import com.digitalasset.ledger.api.v1.ledger_identity_service.LedgerIdentityServiceGrpc.LedgerIdentityService
+import com.digitalasset.ledger.api.v1.testing.time_service.TimeServiceGrpc
+import com.digitalasset.ledger.api.v1.testing.time_service.TimeServiceGrpc.TimeService
+import com.digitalasset.ledger.api.v1.transaction_service.TransactionServiceGrpc
+import com.digitalasset.ledger.api.v1.transaction_service.TransactionServiceGrpc.TransactionService
+import io.grpc.Channel
+
+final class LedgerServices(channel: Channel) {
+  val cmd: CommandService = CommandServiceGrpc.stub(channel)
+  val tx: TransactionService = TransactionServiceGrpc.stub(channel)
+  val id: LedgerIdentityService = LedgerIdentityServiceGrpc.stub(channel)
+  val party: PartyManagementService = PartyManagementServiceGrpc.stub(channel)
+  val time: TimeService = TimeServiceGrpc.stub(channel)
+}
