+++ ledger/ledger-api-test-tool/src/main/scala/com/daml/ledger/acceptance/infrastructure/LedgerTestSuite.scala
@@ -10,7 +10,7 @@ import com.digitalasset.ledger.api.v1.value.{Identifier, Value}
 import com.digitalasset.platform.PlatformApplications
 import com.digitalasset.platform.apitesting.TestTemplateIds
 
-import scala.concurrent.{ExecutionContext, Future}
+import scala.concurrent.Future
 
 private[acceptance] object LedgerTestSuite {
 
@@ -31,6 +31,9 @@ private[acceptance] abstract class LedgerTestSuite(val session: LedgerSession) {
   final def skipIf(reason: String)(p: PartialFunction[LedgerSessionConfiguration, Boolean]) =
     if (p.lift(session.configuration).getOrElse(false)) skip(reason)
 
+  final def applicationId(implicit context: LedgerTestContext): String =
+    context.applicationId
+
   final def offsetAtStart(implicit context: LedgerTestContext): Future[LedgerOffset] =
     context.offsetAtStart
 
@@ -38,25 +41,25 @@ private[acceptance] abstract class LedgerTestSuite(val session: LedgerSession) {
       implicit context: LedgerTestContext): Future[Vector[Transaction]] =
     context.transactionsSinceStart(party, templateIds: _*)
 
-  final def ledgerId()(implicit ec: ExecutionContext): Future[String] =
-    session.ledgerId()
+  final def ledgerId()(implicit context: LedgerTestContext): Future[String] =
+    context.ledgerId()
 
-  final def allocateParty()(implicit ec: ExecutionContext): Future[String] =
-    session.allocateParty()
+  final def allocateParty()(implicit context: LedgerTestContext): Future[String] =
+    context.allocateParty()
 
-  final def allocateParties(n: Int)(implicit ec: ExecutionContext): Future[Vector[String]] =
-    session.allocateParties(n)
+  final def allocateParties(n: Int)(implicit context: LedgerTestContext): Future[Vector[String]] =
+    context.allocateParties(n)
 
   final def create(party: String, templateId: Identifier, args: Map[String, Value.Sum])(
-      implicit ec: ExecutionContext): Future[String] =
-    session.create(party, templateId, args)
+      implicit context: LedgerTestContext): Future[String] =
+    context.create(party, templateId, args)
 
   final def exercise(
       party: String,
       templateId: Identifier,
       contractId: String,
       choice: String,
-      args: Map[String, Value.Sum])(implicit ec: ExecutionContext): Future[Unit] =
-    session.exercise(party, templateId, contractId, choice, args)
+      args: Map[String, Value.Sum])(implicit context: LedgerTestContext): Future[Unit] =
+    context.exercise(party, templateId, contractId, choice, args)
 
 }
