+++ ledger/ledger-api-test-tool/src/main/scala/com/daml/ledger/acceptance/infrastructure/package.scala
@@ -1,0 +1,12 @@
+// Copyright (c) 2019 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package com.daml.ledger.acceptance
+
+import com.digitalasset.ledger.api.v1.ledger_offset.LedgerOffset
+
+package object infrastructure {
+
+  val ledgerEnd = LedgerOffset(LedgerOffset.Value.Boundary(LedgerOffset.LedgerBoundary.LEDGER_END))
+
+}
